refactor: rename file: `results-cache.ts` -> `results-cache-base.ts`
