+++ core/src/plugins/kubernetes/helm/helm-pod.ts
@@ -21,7 +21,7 @@ import { helmPodRunSchema } from "./config.js"
 import { runAndCopy } from "../run.js"
 import { filterManifests, prepareManifests, prepareTemplates } from "./common.js"
 import { kubernetesRunOutputsSchema } from "../kubernetes-type/config.js"
-import { composeCacheableResult, getResultCache, toActionStatus } from "../results-cache.js"
+import { composeCacheableResult, getResultCache, toActionStatus } from "../results-cache-base.js"
 
 const helmRunPodOutputsSchema = kubernetesRunOutputsSchema
 const helmTestPodOutputsSchema = helmRunPodOutputsSchema
