+++ core/src/plugins/kubernetes/results-cache-base.ts
