+++ core/test/integ/src/plugins/kubernetes/kubernetes-type/kubernetes-exec-run.ts
@@ -14,7 +14,7 @@ import type { ConfigGraph } from "../../../../../../src/graph/config-graph.js"
 import { getKubernetesTestGarden } from "./common.js"
 import { RunTask } from "../../../../../../src/tasks/run.js"
 
-import { getResultCache } from "../../../../../../src/plugins/kubernetes/results-cache.js"
+import { getResultCache } from "../../../../../../src/plugins/kubernetes/results-cache-base.js"
 
 describe("kubernetes-type exec Run", () => {
   let garden: TestGarden
